Final Feature Update
As we head closer to the deadline, the final update in terms of features has now been released. Lots of new elements have been implemented, such as a Favicon, animations throughout the site, improved responsivity, and a brand new icon!
Major bugfixes and polishing are here too, such as better refined Grid and Flexbox layouts, a redesign for the new Modal, tweaks to spacing, whitespace, animations, and so much more!